Move vtk ospray classes from VisWindow/Colleagues to Plotter/vtk. (#64)
* Move vtk OSPRay files to appropriate directory.
* Create a namespace for vtkVisItViewNodeFactory.
