+++ src/avt/Plotter/CMakeLists.txt
@@ -128,6 +128,14 @@ vtk/vtkVisItScalarBarWithOpacityActor.C
 vtk/vtkVisItTextActor.C
 )
 
+if(VISIT_OSPRAY)
+    list(APPEND AVTPLOTTER_VTK_SOURCES
+         vtk/vtkOSPRayVisItCubeAxesActorNode.C
+         vtk/vtkOSPRayVisItAxisActorNode.C
+         vtk/vtkVisItViewNodeFactory.C
+        )
+endif()
+
 SET(AVTPLOTTER_SOURCES
 avtActor.C
 avtBehavior.C
@@ -202,12 +210,21 @@ TARGET_LINK_LIBRARIES(avtplotter_ser
     visitcommon avtmath avtview avtfilters_ser avtpipeline_ser visit_vtk 
     lightweight_visit_vtk vtkFiltersHybrid vtkRenderingOpenGL2 
     vtkRenderingFreeType ${OPENGL_LIBRARIES})
+if(VISIT_OSPRAY)
+    target_link_libraries(avtplotter_ser vtkRenderingOSPRay)
+endif()
 VISIT_INSTALL_TARGETS(avtplotter_ser)
 
 #********************************* PARALLEL **********************************
 IF(VISIT_PARALLEL)
     ADD_PARALLEL_LIBRARY(avtplotter_par ${AVTPLOTTER_SOURCES})
-    TARGET_LINK_LIBRARIES(avtplotter_par visitcommon avtmath avtview avtfilters_par avtpipeline_par visit_vtk lightweight_visit_vtk vtkFiltersHybrid vtkRenderingOpenGL2 vtkRenderingFreeType ${OPENGL_LIBRARIES})
+    TARGET_LINK_LIBRARIES(avtplotter_par
+        visitcommon avtmath avtview avtfilters_par avtpipeline_par visit_vtk
+        lightweight_visit_vtk vtkFiltersHybrid vtkRenderingOpenGL2
+        vtkRenderingFreeType ${OPENGL_LIBRARIES})
+    if(VISIT_OSPRAY)
+        target_link_libraries(avtplotter_par vtkRenderingOSPRay)
+    endif()
     VISIT_INSTALL_TARGETS(avtplotter_par)
 ENDIF(VISIT_PARALLEL)
